[Chore] Split haruki_3d.rs into four modules
1 123 lines mixing where the work tree lives, which bundles the export
needs, staging them, invoking the external exporter, and orchestrating
the plan that ties those together.

    paths     the work tree and state files          (leaf)
    tasks     which bundles the export needs         (leaf)
    exporter  invoking it and reading what it said   (leaf)
    staging   staging bundles, cleaning up a failure -> exporter tasks

The facade keeps the five methods that drive a plan, plus the tests.

Pure move: 44 items in, 44 out.

`DownloadTask` widens to `pub(crate)`. It appears in
`append_haruki_3d_download_tasks`, which the download path calls, so the
type has to be at least as visible as the method naming it -- something
the single file never had to state.
